+++ CHANGELOG.md
@@ -276,6 +276,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We fixed an issue where some buttons were not aligned in entry editor. [#16485](https://github.com/JabRef/jabref/issues/16485)
 - We fixed an issue where long recent library paths overflowed and were cut off in the Welcome tab. [#16808](https://github.com/JabRef/jabref/issues/16808)
 - We fixed an issue where a library file containing merge conflict markers was opened without any warning. [#9167](https://github.com/JabRef/jabref/issues/9167)
+- We fixed an issue where the preferences search did not scroll to its first match. [JabRef/jabref-issue-melting-pot#1361](https://github.com/JabRef/jabref-issue-melting-pot/issues/1361)
 
 ### Removed